:construction: Split AppDelegate across files
